1. Add excpetion handler 2. Creating new table for users and role 3. Adding role to database 4. Blocking some elemtns menu for not logged users 5. Prepare application to authorized controllers by rang users

+++ SignLanguage.EF/TableModel/ADOUser.cs
@@ -9,7 +9,7 @@ namespace SignLanguage.EF
         public string IdUser { get; set; }
         public string Login { get; set; }
         public DateTime PasswordExpiredDate { get; set; }
-        public int UserRole { get; set; } // na enuma pozniej
+        public Int16 UserRole { get; set; } // na enuma pozniej
         public string Email { get; set; }
         public bool EmailConfirmed { get; set; }
         public string Password { get; set; }
